FEAT: Start migrating data-related scripts to the models module
